Renamed containing folder to Brawlpy
Small because I forgot to commit this and I don't feel up to working on
anything else for the moment.

The correct name (Brawl.py) would cause issues with the period,
therefore, I removed the period to rectify the issue.

Also changed like one line in the client to make it actually make sense.